Major refactor and API change
I broke up TestCPP.h and the corresponding .cpp file into many pieces.
It just started to get too big, and I wanted more focused TUs.
So this is the first commit for that, where I split them out into
different header files, put them in an 'internal' subdirectory, and
turned TestCPP.h into an aggregate include for including the types and
definitions needed by the library.
So the common definitions are now split out into their own header-only
definition in internal/TestCPPCommon.h.
The exception types are now in their own internal/TestCPPExceptions.h
header with corresponding .cpp.
TestCase and TestSuite now have their own headers,
internal/TestCPPTestCase.h and internal/TestCPPTestSuite.h, with
corresponding .cpps.
And finally TestCPPUtil.h is moved into internal and the TestObjName
type has been moved into the Util header and definition .cpp, but is
still in the base namespace TestCPP, not in TestCPP::Util.

The biggest functional change is splitting the assertions out of
TestSuite and into their own Assertions class, in their own TU.
So from now on, the TestCPP assertions should be referenced through the
Assertions class, not the TestSuite class, which makes a lot more
sense.
I added a couple new assertions a few weeks back, and I can see adding
more, so they really need their own type to encapsulate that
functionality.

There's some significant cleanup to do, but that will probably be in a
separate PR.
